# A few more perf improvements