- Improves the fenced code block parser a bit
- Avoids a dictionary lookup per object when rendering (using linear scans over small arrays instead)
- Misc

| Method         | Job  | Mean       | Error    | Ratio |
|--------------- |----- |-----------:|---------:|------:|
| Parse          | main | 2,512.1 us | 13.63 us |  1.00 |
| Parse          | pr   | 2,417.2 us | 14.36 us |  0.96 |
|                |      |            |          |       |
| ParseAdvanced  | main | 7,872.1 us | 22.51 us |  1.00 |
| ParseAdvanced  | pr   | 7,906.6 us | 44.54 us |  1.00 |
|                |      |            |          |       |
| Render         | main |   716.3 us |  2.38 us |  1.00 |
| Render         | pr   |   688.6 us |  4.58 us |  0.96 |
|                |      |            |          |       |
| RenderAdvanced | main | 1,073.6 us |  4.24 us |  1.00 |
| RenderAdvanced | pr   |   984.2 us |  6.25 us |  0.92 |